new files:
radial_interpolation.[cxx,h]
provides 1D Interpolation functions (so far pcw. linear and pcw. quadratic are provided)
MISSING: implementation of RI for a set of directions (but straight-forward)

tangential_interpolation.[cxx,h]
provides the kernel based tangential interpolation (without derivatives, so far)
and it comprises an export of the weights for the individual directions (for use with RI)